refactor(dto): split DTOs into Input and Output
