+++ src/Dto/Input/ImportGitHubEventsInput.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Dto;
+namespace App\Dto\Input;
 
 class ImportGitHubEventsInput
 {
