+++ src/Repository/ReadEventRepository.php
@@ -2,7 +2,7 @@
 
 namespace App\Repository;
 
-use App\Dto\SearchInput;
+use App\Dto\Input\SearchInput;
 
 interface ReadEventRepository
 {
